Implement movie detail caching with Room and refactor network handling to use Kotlin's `Result`.
- **Caching & Persistence**:
    - Implement `CacheDao` and `CacheRepository` to manage local storage for movie details, genres, credits, and videos.
    - Define Room entities and cross-reference tables for `MovieDetail` and its nested components (Cast, Crew, Genres, etc.).
    - Update `AppDatabase` to include new entities and export the schema.
    - Implement `isMovieComplete` logic to determine if cached data satisfies the required `AppendedToResponse` types.

- **Data Layer Refactoring**:
    - Replace custom `NetworkResult` with Kotlin's native `Result` across repositories and use cases.
    - Refactor `TMDBRepositoryImpl` to integrate with `CacheRepository`, providing a single source of truth for movie details.
    - Update `Genre`, `ProductionCompany`, and `MediaItemDTO` to use `Int` instead of `Long` for IDs for consistency with the database layer.
    - Rename mapping functions from `mapToDomain()` to `toDomain()` for brevity.

- **Logic & ViewModel Updates**:
    - Update `MovieViewModel` and `ProfileViewModel` to use `flatMapLatest` and `flow` for reactive data fetching with `Result` handling.
    - Adjust `buildAppendQuery` and its usages to accept a `List` instead of `vararg` for better type safety and flexibility.
    - Refine `safeApiCall` and `onSuccess` extensions to support the transition to `Result`.

- **Dependency Injection**:
    - Bind `CacheRepository` and provide `CacheDao` in `DatabaseModule` and `RepositoryModule`.

+++ feature/tmdb/src/main/java/com/vroff/tmdb/entity/search/MediaItemDTO.kt
@@ -26,7 +26,7 @@ data class MediaItemDTO(
     @SerializedName("original_language")
     val originalLanguage: String?,
     @SerializedName("genre_ids")
-    val genreIds: List<Long>?,
+    val genreIds: List<Int>?,
     val popularity: Double,
     @SerializedName("first_air_date")
     val firstAirDate: String?,
@@ -51,7 +51,7 @@ data class MediaItemDTO(
     val video: Boolean?,
 ) {
     fun mapToDomain(
-        genreMapper: (List<Long>?, MediaType) -> List<Genre>,
+        genreMapper: (List<Int>?, MediaType) -> List<Genre>,
         selectedMediaType: MediaType? = null,
     ): TMDBMediaItem {
         val mediaType = selectedMediaType ?: MediaType.entries.associateBy { it.type }[mediaType] ?: MediaType.UNKNOWN
